se corrije error producido en la consulta de eliminar el registro en el modelo de BD Pelicula

+++ peliculas/models.py
@@ -13,5 +13,6 @@ class Pelicula(models.Model):
     return fila
 
   def delete(self, using=None, keep_parents=False):
-    self.imagen.storage.delete(self.imagen.name)
-    super().delete()
+    if self.imagen:
+      self.imagen.storage.delete(self.imagen.name)
+    super().delete(using=using, keep_parents=keep_parents)
